requirements.txt
tabulate>=0.9.0
python-docx>=1.1.0
